Check if there are other boot_info_script.sh files in the same directory.
Check if there are other boot_info_script.sh files in the same directory.
Display all BIS files found in that directory.

To force running a BIS file, add "--this" as first argument to the script.

This can be useful when BIS was downloaded multiple times with Firefox, Chromium, ...
Those browsers will add a suffix to the filename, when there was already
a file with the same name:

   boot_info_script(<number>).sh

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>