Only require "unlzma" when Grub2 (v1.99) is detected.
Only require "unlzma" when Grub2 (v1.99) is detected.
"unlzma" is needed for getting the core_dir of Grub2 (v1.99), but isn't needed
for lower versions of Grub2.

When unlzma isn't installed, and when Grub2 (v1.99) is detected, the other
info can still be shown. A warning about missing "unlzma" will be added at
the end of the output file.

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>