Also check for awk support in the "/usr/lib/initramfs-tools/bin/busybox" binary.
Not all Ubuntu versions have busybox installed by default in the $PATH.
Ubuntu 8.04 has a busybox version in "/usr/lib/initramfs-tools/bin/busybox"
that seems to work reasonably well in most cases, so installing gawk isn't
always needed.
In at least Ubuntu 10.10, "/usr/lib/initramfs-tools/bin/busybox" is a small
busybox version without "Math support is not compiled in" and doesn't work
in combination with BIS.
The default busybox version in the $PATH, is a "fat" busybox binary, which
works well.
So we first check for the presence of busybox in the $PATH before trying
"/usr/lib/initramfs-tools/bin/busybox".

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>