Get exact Syslinux version before checking the integrity in syslinux_info function.
Get exact Syslinux version before checking the integrity of Syslinux
in the syslinux_info function.
When LDLINUX_MAGIC is found in the second sector, the Syslinux version
string will be intact too in most cases (located before the LDLINUX_MAGIC).
So even if the integrity check of Syslinux fails, we can display a more
exact Syslinux version number.

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>