Add support for displaying the ADV info to the syslinux_info function.
Add support for displaying the ADV info to the syslinux_info function.

The ADV tells Syslinux to, for example boot a certain label only one time
(useful to test a new kernel).
The ADV is also used for overruling the default label set in the
Syslinux config file. When MENU SAVE is available in the Syslinux
config file, the entry that was booted will become the new default
label.

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>