Add a CHANGELOG.
Add a CHANGELOG.

Diffing was done between older versions of boot_info_script to find the
changes made between those versions. No CHANGELOG was made in the past,
so the reason for some changes in the source code are based on the code
itself. So mistakes in the listed changes are possible.

Signed-off-by: Gert Hulselmans <ghulselmans@users.sourceforge.net>