Send informative error messages to stderr instead of to Error_Log file directly.
Send informative error messages to stderr instead of to Error_Log file directly.

Appending messages to the Error_Log file doesn't work on at least Fedora,
when stderr is redirected to a file with:

  exec 2> ${Error_Log};

So it is better to write the messages to stderr instead of appending them to
the Error_Log file directly.