Fix crash of BIS when there is no embedded menu in grub4dos/wee.
Fix crash of BIS when there is no embedded menu in grub4dos/wee.
When no offset is found for the embedded menu, the ${offset_menu} variable
is empty. The check to see if the variable was empty was missing quotes,
which made the check always return true even if the variable was empty.

This patch fixes the folowing error message:
  (  / 2 ) + 16 : syntax error: operand expected (error token is "/ 2 ) + 16 ")