Remove padding bytes in embedded config
GRUB does not fill bytes resulting from alignment, except for the
initial '\0' character. BIS pattern did not catch it then; '\0' was lost
during variable assignment and BIS displayed garbage at the end. So delete
everything starting with '\0', do not expect to have zeroes only.

For prefix the whole memory buffer was inittialized with zeroes and this
did not happen.